[wip] stt
