--- lightning_transformers/task/nlp/asr/__init__.py
@@ -1,3 +1,0 @@
-from lightning_transformers.task.vision.image_classification.config import ImageClassificationDataConfig  # noqa: F401
-from lightning_transformers.task.vision.image_classification.data import ImageClassificationDataModule  # noqa: F401
-from lightning_transformers.task.vision.image_classification.model import ImageClassificationTransformer  # noqa: F401
