--- lightning_transformers/task/nlp/asr/config.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-from lightning_transformers.core.config import TransformerDataConfig
-
-
-@dataclass
-class SpeechRecognitionDataConfig(TransformerDataConfig):
-    ...
